fix: update conditional to always check for id existence
