+++ src/hooks/use-update-distribution-list.test.ts
@@ -6,79 +6,12 @@
 
 import { faker } from '@faker-js/faker';
 
-import { getMembersPartition, useUpdateDistributionList } from './use-update-distribution-list';
+import { useUpdateDistributionList } from './use-update-distribution-list';
 import { screen, setupHook } from '../carbonio-ui-commons/test/test-setup';
 import { JEST_MOCKED_ERROR } from '../constants/tests';
 import { DistributionList } from '../model/distribution-list';
 import { registerDistributionListActionHandler } from '../tests/msw-handlers/distribution-list-action';
 import { generateDistributionList } from '../tests/utils';
-
-describe('getMembersPartition', () => {
-	it("should return an object with empty members if the two parameters don't contain any value", () => {
-		const originalMembers: Array<string> = [];
-		const updatedMembers: Array<string> = [];
-		const result = getMembersPartition(originalMembers, updatedMembers);
-		expect(result).toEqual({ membersToAdd: [], membersToRemove: [] });
-	});
-
-	it('should return an object with empty members if the two parameters contain the same values', () => {
-		const originalMembers: Array<string> = ['john.doe@domain.org', 'mario.rossi@dominio.it'];
-		const updatedMembers: Array<string> = ['mario.rossi@dominio.it', 'john.doe@domain.org'];
-		const result = getMembersPartition(originalMembers, updatedMembers);
-		expect(result).toEqual({ membersToAdd: [], membersToRemove: [] });
-	});
-
-	it('should return an object with 3 members to add if the second parameter contains 3 members not included in the first parameter', () => {
-		const originalMembers: Array<string> = ['john.doe@domain.org', 'mario.rossi@dominio.it'];
-		const updatedMembers: Array<string> = [
-			'mario.rossi@dominio.it',
-			'john.doe@domain.org',
-			'1@dominio.it',
-			'2@dominio.it',
-			'3@dominio.it'
-		];
-		const result = getMembersPartition(originalMembers, updatedMembers);
-		expect(result).toEqual({
-			membersToAdd: ['1@dominio.it', '2@dominio.it', '3@dominio.it'],
-			membersToRemove: []
-		});
-	});
-
-	it('should return an object with 3 members to remove if the first parameter contains 3 members not included in the second parameter', () => {
-		const originalMembers: Array<string> = [
-			'mario.rossi@dominio.it',
-			'john.doe@domain.org',
-			'1@dominio.it',
-			'2@dominio.it',
-			'3@dominio.it'
-		];
-		const updatedMembers: Array<string> = ['john.doe@domain.org', 'mario.rossi@dominio.it'];
-		const result = getMembersPartition(originalMembers, updatedMembers);
-		expect(result).toEqual({
-			membersToAdd: [],
-			membersToRemove: ['1@dominio.it', '2@dominio.it', '3@dominio.it']
-		});
-	});
-
-	it('should return an object with 2 members to add and 3 members to remove if the parameters contains different members', () => {
-		const originalMembers: Array<string> = [
-			'1@dominio.it',
-			'2@dominio.it',
-			'anonymous@hacker.net',
-			'3@dominio.it'
-		];
-		const updatedMembers: Array<string> = [
-			'john.doe@domain.org',
-			'anonymous@hacker.net',
-			'mario.rossi@dominio.it'
-		];
-		const result = getMembersPartition(originalMembers, updatedMembers);
-		expect(result).toEqual({
-			membersToAdd: ['john.doe@domain.org', 'mario.rossi@dominio.it'],
-			membersToRemove: ['1@dominio.it', '2@dominio.it', '3@dominio.it']
-		});
-	});
-});
 
 describe('Use update distribution list hook', () => {
 	it('should show success snackbar when there are no errors', async () => {
